pyproject.toml
[project]
name = "researchproject1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "notebook>=7.5.1",
    "numpy>=2.4.0",
    "pandas>=2.1.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "librosa>=0.11.0",
    "tensorflow>=2.20.0",
    "matplotlib>=3.10.0",
    "soundfile>=0.12.1",
    "pillow>=10.0.0",
    "python-multipart>=0.0.9",
    "scikit-learn>=1.5.0",
    "joblib>=1.4.0",
    "ipyleaflet>=0.19.0",
    "ipywidgets>=8.1.0",
]
